Extract persistence related helpers to their own file
