+++ lib/helpers.ts
@@ -9,9 +9,9 @@ import { RocketChatAssociationModel, RocketChatAssociationRecord } from '@rocket
 import { IRoom } from '@rocket.chat/apps-engine/definition/rooms';
 import { IUser } from '@rocket.chat/apps-engine/definition/users';
 
-import { AppSetting } from './app-settings';
-import { AppInfoEnum } from './enums/AppInfoEnum';
-import * as jwt from './jwt';
+import { AppSetting } from '../app-settings';
+import { AppInfoEnum } from '../enums/AppInfoEnum';
+import * as jwt from '../jwt';
 
 export async function startNewMessageWithDefaultSenderConfig(modify: IModify, read: IRead, sender: IUser, room?: IRoom): Promise<IMessageBuilder> {
     const settingsReader = read.getEnvironmentReader().getSettings();
@@ -58,28 +58,4 @@ export async function getUrlAndAuthToken(read: IRead, path: string, method: stri
         url: `${authData.baseUrl}${path}`,
         token,
     };
-}
-
-interface IConnectedProjectsRecord {
-    room: string;
-    connectedProjects: object;
-}
-
-export async function getConnectedProjects(persistence: IPersistenceRead, room?: IRoom): Promise<Array<IConnectedProjectsRecord>> {
-    const associations = [new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, 'projects')];
-
-    if (room) {
-        associations.push(new RocketChatAssociationRecord(RocketChatAssociationModel.ROOM, room.id));
-    }
-
-    return await persistence.readByAssociations(associations) as Array<IConnectedProjectsRecord>;
-}
-
-export async function persistConnectedProjects(persis: IPersistence, room: IRoom, connectedProjects: object): Promise<void> {
-    const roomAssociation = new RocketChatAssociationRecord(RocketChatAssociationModel.ROOM, room.id);
-    const projectsAssociation = new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, 'projects');
-
-    await persis.removeByAssociations([roomAssociation, projectsAssociation]);
-
-    await persis.createWithAssociations({ room: room.id, connectedProjects }, [roomAssociation, projectsAssociation]);
 }
