+++ onComment.ts
@@ -11,7 +11,7 @@ import { ApiEndpoint } from '@rocket.chat/apps-engine/definition/api/ApiEndpoint
 import { IApiEndpointInfo } from '@rocket.chat/apps-engine/definition/api/IApiEndpointInfo';
 
 import { WebhookEventEnum } from './enums/WebhookEventEnum';
-import { parseJiraDomainFromIssueUrl, startNewMessageWithDefaultSenderConfig } from './helpers';
+import { parseJiraDomainFromIssueUrl, startNewMessageWithDefaultSenderConfig } from './lib/helpers';
 
 export class OnCommentEndpoint extends ApiEndpoint {
     public path: string = 'on_comment';
