+++ sdk/index.ts
@@ -1,5 +1,5 @@
 import { IHttp, IHttpResponse, IRead } from '@rocket.chat/apps-engine/definition/accessors';
-import { getUrlAndAuthToken } from '../helpers';
+import { getUrlAndAuthToken } from '../lib/helpers';
 
 export interface IAvatarUrls {
     '48x48': string;
